HeadHunter--T/requirements.txt
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
h5py==3.2.1
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.4.2
networkx==2.5.1
numpy==1.20.3
ordered-set==4.0.2
pillow==8.2.0
plyfile==0.7.4
pyparsing==2.4.7
python-dateutil==2.8.1
pywavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.6.3
six==1.16.0
tifffile==2021.4.8
torchmeta==1.7.0
tqdm==4.61.0
trimesh==3.9.19
typing-extensions==3.10.0.0
urllib3==1.26.5
opencv-python
pyyaml
albumentations
easydict
pycocotools
munkres
